requirements.txt
apify_client==1.6.2
apify_shared==1.1.1
pandas==2.2.0
psycopg2-binary==2.9.9
pyarrow==15.0.0
python-dotenv==1.0.1
requests==2.31.0
SQLAlchemy==2.0.25
